[ refactor ] Move constructors from `ast` to `ast_cons`
