+++ src/syntax/core/mod.rs
@@ -1,10 +1,12 @@
 use super::common::{SyntaxInfo, ToSyntaxInfo};
 
 pub use self::ast::*;
+pub use self::ast_cons::*;
 pub use self::level::*;
 pub use self::pretty::*;
 
 mod ast;
+mod ast_cons;
 mod level;
 mod pretty;
